feat: rename app to Collect :fire:
