+++ src/index.jsx
@@ -5,7 +5,7 @@ import { render } from 'react-dom'
 import { Router, Route, Redirect, hashHistory } from 'react-router'
 
 import { I18n } from './plugins/i18n'
-import DataConnectStore, { Provider } from './lib/DataConnectStore'
+import CollectStore, { Provider } from './lib/CollectStore'
 
 import App from './containers/App'
 import DiscoveryList from './components/DiscoveryList'
@@ -38,7 +38,7 @@ document.addEventListener('DOMContentLoaded', () => {
   window.initKonnectors = require('./initKonnectors.json')
   window.initFolders = require('./initFolders.json')
 
-  const store = new DataConnectStore(window.initKonnectors, window.initFolders, context)
+  const store = new CollectStore(window.initKonnectors, window.initFolders, context)
   const useCases = store.getUseCases()
 
   render((
